feat(AED): done things on the jpa project and an error on the pgl project

+++ AED/java/InstitutoJPA/src/main/java/es/iespto/angel/InstitutoJPA/entity/Alumno.java
@@ -24,7 +24,7 @@ public class Alumno implements Serializable {
 	@Column(length=50)
 	private String apellidos;
 
-	private BigInteger fechanacimiento;
+	private Long fechanacimiento;
 
 	@Column(length=50)
 	private String nombre;
@@ -52,11 +52,11 @@ public class Alumno implements Serializable {
 		this.apellidos = apellidos;
 	}
 
-	public BigInteger getFechanacimiento() {
+	public Long getFechanacimiento() {
 		return this.fechanacimiento;
 	}
 
-	public void setFechanacimiento(BigInteger fechanacimiento) {
+	public void setFechanacimiento(Long fechanacimiento) {
 		this.fechanacimiento = fechanacimiento;
 	}
